database refactored, added JS Datatables, mistakes fixed, added some controllers

+++ StrongmanApp/Controllers/CountriesController.cs
@@ -2,9 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+//using StrongmanApp.Context;
 using StrongmanApp.Models;
 
 namespace StrongmanApp.Controllers
@@ -41,13 +43,13 @@ namespace StrongmanApp.Controllers
 
             return View(country);
         }
-
+        [Authorize(Roles = "admin")]
         // GET: Countries/Create
         public IActionResult Create()
         {
             return View();
         }
-
+        [Authorize(Roles = "admin")]
         // POST: Countries/Create
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
@@ -55,6 +57,10 @@ namespace StrongmanApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name,Details,FlagUrl")] Country country)
         {
+            if (_context.Countries.Where(a => a.Name == country.Name).Count() > 0)
+            {
+                ModelState.AddModelError("Name", "The name already exists");
+            }
             if (ModelState.IsValid)
             {
                 _context.Add(country);
@@ -63,7 +69,7 @@ namespace StrongmanApp.Controllers
             }
             return View(country);
         }
-
+        [Authorize(Roles = "admin")]
         // GET: Countries/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -79,7 +85,7 @@ namespace StrongmanApp.Controllers
             }
             return View(country);
         }
-
+        [Authorize(Roles = "admin")]
         // POST: Countries/Edit/5
         // To protect from overposting attacks, enable the specific properties you want to bind to.
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
@@ -91,7 +97,10 @@ namespace StrongmanApp.Controllers
             {
                 return NotFound();
             }
-
+            if (_context.Countries.Where(a => a.Name == country.Name).Count() > 0)
+            {
+                ModelState.AddModelError("Name", "The name already exists");
+            }
             if (ModelState.IsValid)
             {
                 try
@@ -114,7 +123,7 @@ namespace StrongmanApp.Controllers
             }
             return View(country);
         }
-
+        [Authorize(Roles = "admin")]
         // GET: Countries/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
@@ -132,7 +141,7 @@ namespace StrongmanApp.Controllers
 
             return View(country);
         }
-
+        [Authorize(Roles = "admin")]
         // POST: Countries/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
@@ -141,7 +150,15 @@ namespace StrongmanApp.Controllers
             var country = await _context.Countries.FindAsync(id);
             if (country != null)
             {
-                _context.Countries.Remove(country);
+                if (_context.Towns.Where(a => a.CountryId == id).Count() > 0) {
+                    ModelState.AddModelError("Name", "This country has connected cities");
+                    return (View(country));
+                }
+                else
+                {
+                    _context.Countries.Remove(country);
+                }
+               
             }
 
             await _context.SaveChangesAsync();
